Fix cronjob errors of Dovecot expunge
Recently, I discovered that the cronjob on my mailcow-dockerized host isn't working.
If the script is called via crontab, the docker-compose command doesn't know where the config file is located. So I added the "cd" to change the directory to the working directory of the mailcow-dockerized instance.
When executing the docker-compose exec command via cronjob, it fails with error message "the input device is not a TTY". By adding the parameter "-T" it works.