Update prerequisite-dns.md
I have had issues with mail delivery to some mailservers until I added this registry: `mail IN TXT "v=spf1 a ~all"`. May be you want to suggest it in docs.
Thank you for such an awesome project. I started to use it two days ago and I have fallen in love with it.